+++ node/node_workpackage_test.go
@@ -130,8 +130,8 @@ func TestAvailabilityReconstruction(t *testing.T) {
 		t.Fatalf("Error setting up nodes: %v\n", err)
 	}
 	nodes := make([]*Node, numNodes)
-	for i := 0; i < numNodes; i++ {
-		node, err := newNode(uint32(i), validatorSecrets[i], &genesisConfig, peers, peerList, DAFlag, nodePaths[i], basePort+i)
+	for i := uint16(0); i < numNodes; i++ {
+		node, err := newNode(i, validatorSecrets[i], &genesisConfig, peers, peerList, DAFlag, nodePaths[i], int(basePort+i))
 		if err != nil {
 			t.Fatalf("Failed to create node %d: %v\n", i, err)
 		}
